requirements.txt
flask
Flask-SQLAlchemy
flask-login
Flask-Paginate
alembic
Flask-WTF
email_validator